added test for cli.py serialize
